fix: build errors on question creation pages and questions discovery
