+++ front_end/src/app/(main)/questions/create/question/page.tsx
@@ -59,4 +59,4 @@ const QuestionCreator: React.FC<{ searchParams: SearchParams }> = async ({
   );
 };
 
-export default WithServerComponentErrorBoundary(QuestionCreator);
+export default QuestionCreator;
